--- docs/source/reference/api/eland.ml.ImportedMLModel.predict.rst
@@ -1,6 +1,0 @@
-eland.ml.ImportedMLModel.predict
-================================
-
-.. currentmodule:: eland.ml
-
-.. automethod:: ImportedMLModel.predict
