--- docs/source/reference/api/eland.ml.ImportedMLModel.rst
@@ -1,6 +1,0 @@
-eland.ml.ImportedMLModel
-========================
-
-.. currentmodule:: eland.ml
-
-.. autoclass:: ImportedMLModel
