+++ docs/source/reference/api/eland.ml.MLModel.import_model.rst
@@ -1,0 +1,6 @@
+eland.ml.MLModel.import_model
+=============================
+
+.. currentmodule:: eland.ml
+
+.. automethod:: MLModel.import_model
