+++ docs/source/reference/api/eland.ml.MLModel.predict.rst
@@ -1,0 +1,6 @@
+eland.ml.MLModel.predict
+========================
+
+.. currentmodule:: eland.ml
+
+.. automethod:: MLModel.predict
