+++ eland/ml/common.py
@@ -15,9 +15,5 @@
 #  specific language governing permissions and limitations
 #  under the License.
 
-from eland.ml.ml_model import MLModel, ImportedMLModel
-
-__all__ = [
-    "MLModel",
-    "ImportedMLModel",
-]
+TYPE_CLASSIFICATION = "classification"
+TYPE_REGRESSION = "regression"
